requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9